test: move isErfolgAnalog test to main file
